pyproject.toml
[tool.poetry]
name = "repl-nix-ssimvisualizer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.38.0"
matplotlib = "^3.9.2"
scikit-image = "^0.24.0"
numpy = "^2.1.1"
pillow = "^10.4.0"
pyqt5 = "^5.15.11"
opencv-python = "^4.10.0.84"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
